+++ frontend/src/pages/home.js
@@ -5,6 +5,7 @@ import { userGetUserID } from '../hooks/useGetUserID';
 import axios from 'axios';
 import GroceryList from '../components/grocerylist';
 import './home.css';
+import { useGroceryList, GroceryListProvider } from '../components/grocerylist-context';
 
 
 export const Home = () => {
